requirements.txt
numpy
pandas
pyyaml
scikit-learn
scikit-image
nibabel
torch
monai
matplotlib
seaborn
plotly
tqdm
